+++ app/src/main/res/xml/recipe_widget_provider_info.xml
@@ -6,5 +6,5 @@
     android:minWidth="170dp"
     android:previewImage="@drawable/ic_local_dining_black_24dp"
     android:resizeMode="horizontal|vertical"
-    android:updatePeriodMillis="86400000"
+    android:updatePeriodMillis="0"
     android:widgetCategory="home_screen" />
